# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/riot/auth/submit-url": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Принимает URL, скопированный из браузера после авторизации через Riot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Завершить OAuth-привязку через вставку URL",
                "parameters": [
                    {
                        "description": "URL из браузера после авторизации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riot.submitRedirectURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аккаунт успешно привязан",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос или токен не найден",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    }
                }
            }
        },
        "/v1/riot/auth/url": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Генерирует OAuth URL на страницу входа Riot. Откройте эту ссылку в браузере,",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Получить ссылку для авторизации через Riot",
                "responses": {
                    "200": {
                        "description": "OAuth ссылка сформирована",
                        "schema": {
                            "$ref": "#/definitions/riot.authURLResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка генерации ссылки",
                        "schema": {
                            "$ref": "#/definitions/riot.authURLResponse"
                        }
                    }
                }
            }
        },
        "/v1/riot/callback": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Принимает токены от клиентского OAuth-flow и привязывает аккаунт",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Привязка Riot аккаунта через OAuth",
                "parameters": [
                    {
                        "description": "OAuth токены",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riot.riotCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аккаунт успешно привязан",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    }
                }
            }
        },
        "/v1/riot/login": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Авторизует пользователя на серверах Riot по логину и паролю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Прямой вход через Riot аккаунт",
                "parameters": [
                    {
                        "description": "Данные входа в Riot",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Требуется капча или 2FA",
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    }
                }
            }
        },
        "/v1/riot/login/captcha": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Повторяет авторизацию с решенной hcaptcha",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Отправка решения капчи для входа в Riot",
                "parameters": [
                    {
                        "description": "session_id, учетные данные и токен hcaptcha",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginCaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Требуется 2FA после капчи",
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "401": {
                        "description": "Сессия не найдена или истекла",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    }
                }
            }
        },
        "/v1/riot/login/mfa": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Принимает код из email/authenticator для завершения Riot-авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Riot"
                ],
                "summary": "Подтверждение 2FA кода для входа в Riot",
                "parameters": [
                    {
                        "description": "session_id из предыдущего шага и код 2FA",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginMFARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Неверный код - повторите попытку",
                        "schema": {
                            "$ref": "#/definitions/riot.riotLoginChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "401": {
                        "description": "Сессия не найдена или истекла",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/riot.linkRiotAccountResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "message": {
                    "type": "string"
                },
                "meta": {}
            }
        },
...
        "riot.authURLResponse": {
            "type": "object",
            "properties": {
                "auth_url": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "riot.linkRiotAccountResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "puuid": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "riot.riotCallbackRequest": {
            "type": "object",
            "required": [
                "access_token"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "id_token": {
                    "type": "string"
                }
            }
        },
        "riot.riotLoginCaptchaRequest": {
            "type": "object",
            "required": [
                "captcha_token",
                "password",
                "session_id",
                "username"
            ],
            "properties": {
                "captcha_token": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "riot.riotLoginChallengeResponse": {
            "type": "object",
            "properties": {
                "code_length": {
                    "type": "integer"
                },
                "email": {
                    "description": "MFA",
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/response.ErrorResponse"
                },
                "hcaptcha_key": {
                    "description": "Captcha",
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "riot.riotLoginMFARequest": {
            "type": "object",
            "required": [
                "code",
                "session_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "riot.riotLoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "riot.submitRedirectURLRequest": {
            "type": "object",
            "required": [
                "redirect_url"
            ],
            "properties": {
                "redirect_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}